Added mutex in Range class and found THE BUG
Whenever we started a transaction, we sent a range size.
The transaction now asks for a fixed amount of data.
Meanwhile the range size was changed due to split.
When transaction is ready, it again updated the range size to
the total bytes it had requsted and hence THE BUG was conceived.